+++ semcore/utils/src/addonText.tsx
@@ -4,7 +4,11 @@ export default function addonText(
   children: React.ReactNode,
   Text: React.ComponentType<any>,
   Addon: React.ComponentType,
+  isTextWrapIfEmptyChildren: boolean = false,
 ) {
+  if (isTextWrapIfEmptyChildren && !children) {
+    return false;
+  }
   return React.Children.toArray(children).some(
     (element) =>
       React.isValidElement(element) &&